requirements.txt
beautifulsoup4
requests
transformers
transformers-adapters
sentence-transformers
scikit-learn
numpy
selenium
pandas
tqdm
python-dotenv
